+++ backend/src/interfaces/Job.ts
@@ -24,7 +24,7 @@ export interface BulkLoadJob extends CommonJobData {
 export interface ExportJob extends CommonJobData {
   filter_id: string;
   format: string;
-  dataset_slugs: string[];
+  dataset_ids: string[];
   total_records_estimate: number;
   current_cursor: string | null;
   total_records_processed: number;