+++ backend/src/services/UnitConversionService.ts
@@ -20,16 +20,17 @@ export default class UnitConversionService {
     // Find all unit conversions for the soil property
     const unitConversions = await repo.find({
       where: whereConditions,
+      relations: ['soil_property'],
     });
 
     return unitConversions;
   };
 
   getUnitConversion = async (requestData: RequestData, slug: string): Promise<UnitConversionEntity> => {
-    return await getEntity(requestData, UnitConversionEntity, EntityType.UNIT_CONVERSION, slug);
+    return await getEntity(requestData, UnitConversionEntity, EntityType.UNIT_CONVERSION, slug, ['soil_property']);
   };
 
   getUnitConversionsBySlug = async (requestData: RequestData, slugs: string[]): Promise<UnitConversionEntity[]> => {
-    return await getEntities(requestData, UnitConversionEntity, EntityType.UNIT_CONVERSION, slugs);
+    return await getEntities(requestData, UnitConversionEntity, EntityType.UNIT_CONVERSION, slugs, ['soil_property']);
   };
 }