+++ frontend/soilhive-app/src/components/DownloadPreview/DownloadPreviewFilters/DownloadPreviewFilters.tsx
@@ -11,7 +11,7 @@ import type { PreviewFilters } from 'types/downloadPreview';
 
 function DownloadPreviewFilters({
   soilProperties = [],
-  filters = {},
+  filters = { soil_properties: [] },
   onFiltersChange,
   dialogOpen = false,
   setDialogOpen,
@@ -20,6 +20,7 @@ function DownloadPreviewFilters({
   calendarMinMaxRange = [undefined, undefined],
   depthMinMaxRange = [undefined, undefined],
   fixedDepthRange = null,
+  selectedDatasets,
   onDatasetsChange,
   isLoading = true,
 }: {
@@ -33,11 +34,11 @@ function DownloadPreviewFilters({
   fixedCalendarRange?: Nullable<Array<Date | null>>;
   depthMinMaxRange?: [number | undefined, number | undefined];
   fixedDepthRange?: Nullable<[number, number]>;
-  onDatasetsChange?: (datasets: string[] | undefined) => void;
+  selectedDatasets?: string[];
+  onDatasetsChange?: (datasets: string[]) => void;
   isLoading?: boolean;
 }) {
   const { isMobileLayout } = useDevice();
-  const [selectedDataset, setSelectedDataset] = useState<string>();
   const [selectedDateRange, setSelectedDateRange] = useState<Nullable<Array<Date | null>>>(fixedCalendarRange);
 
   useEffect(() => {
@@ -69,20 +70,14 @@ function DownloadPreviewFilters({
       : undefined;
 
   const onDatasetsDropdownChange = (e: DropdownChangeEvent) => {
-    setSelectedDataset(e.value);
-    onDatasetsChange?.(e.value ? [e.value] : undefined);
+    onDatasetsChange?.([e.value]);
   };
 
   const onSoilPropertiesDropdownChange = (e: DropdownChangeEvent) => {
-    if (e.value) {
-      onFiltersChange?.({
-        ...filters,
-        soil_properties: [e.value],
-      });
-    } else {
-      const { soil_properties: _, ...filtersWithoutSoilProperties } = filters;
-      onFiltersChange?.(filtersWithoutSoilProperties);
-    }
+    onFiltersChange?.({
+      ...filters,
+      soil_properties: [e.value],
+    });
   };
 
   const onDepthRangeDropdownChange = (e: DropdownChangeEvent) => {
@@ -128,9 +123,8 @@ function DownloadPreviewFilters({
       <Dropdown
         className={styles.Dropdown}
         panelClassName={styles.DropdownPanel}
-        value={selectedDataset}
+        value={selectedDatasets?.[0]}
         onChange={onDatasetsDropdownChange}
-        showClear
         options={datasets}
         optionValue="id"
         optionLabel="name"
@@ -140,9 +134,8 @@ function DownloadPreviewFilters({
       <Dropdown
         className={styles.Dropdown}
         panelClassName={styles.DropdownPanel}
-        value={filters.soil_properties?.[0]}
+        value={filters.soil_properties[0]}
         onChange={onSoilPropertiesDropdownChange}
-        showClear
         options={soilProperties}
         optionLabel="property_name"
         optionValue="id"